+++ packages/language-core/lib/types.ts
@@ -12,8 +12,14 @@ export { VueEmbeddedCode };
 export type RawVueCompilerOptions = Partial<Omit<VueCompilerOptions, 'target' | 'plugins'>> & {
 	strictTemplates?: boolean;
 	target?: 'auto' | 3 | 3.3 | 3.5 | 3.6 | 99 | number;
-	plugins?: string[];
+	plugins?: RawPlugin[];
 };
+
+export type RawPlugin =
+	| string
+	| Record<string, any> & {
+		name: string;
+	};
 
 export interface VueCodeInformation extends CodeInformation {
 	__importCompletion?: boolean;
@@ -109,14 +115,16 @@ export interface VueLanguagePluginReturn {
 	resolveEmbeddedCode?(fileName: string, sfc: Sfc, embeddedFile: VueEmbeddedCode): void;
 }
 
-export type VueLanguagePlugin = (ctx: {
-	modules: {
-		typescript: typeof ts;
-		'@vue/compiler-dom': typeof CompilerDOM;
-	};
-	compilerOptions: ts.CompilerOptions;
-	vueCompilerOptions: VueCompilerOptions;
-}) => VueLanguagePluginReturn | VueLanguagePluginReturn[];
+export type VueLanguagePlugin = (
+	ctx: Record<string, any> & {
+		modules: {
+			typescript: typeof ts;
+			'@vue/compiler-dom': typeof CompilerDOM;
+		};
+		compilerOptions: ts.CompilerOptions;
+		vueCompilerOptions: VueCompilerOptions;
+	},
+) => VueLanguagePluginReturn | VueLanguagePluginReturn[];
 
 export interface SfcBlock {
 	name: string;